+++ ui/guest/src/store/epics/root.ts
@@ -101,6 +101,7 @@ import { uploadDataUrl } from '@craftercms/studio-ui/services/content';
 import { getRequestForgeryToken } from '@craftercms/studio-ui/utils/auth';
 import { ensureSingleSlash } from '@craftercms/studio-ui/utils/string';
 import { getInheritanceParentIdsForField } from '@craftercms/studio-ui/utils/content';
+import { SearchItem } from '@craftercms/studio-ui/models';
 
 const createReader$ = (file: File) =>
   new Observable((subscriber: Subscriber<ProgressEvent<FileReader>>) => {
@@ -125,8 +126,7 @@ const epic = combineEpics<GuestStandardAction, GuestStandardAction, GuestState>(
     action$.pipe(
       ofType('mouseover', 'mouseleave'),
       withLatestFrom(state$),
-      filter((args) => args[1].status === EditingStatus.LISTENING),
-      tap(([action]) => action.payload.event.stopPropagation()),
+      tap(([action, state]) => state.status === EditingStatus.LISTENING && action.payload.event.stopPropagation()),
       ignoreElements()
     ),
   // endregion
@@ -260,7 +260,7 @@ const epic = combineEpics<GuestStandardAction, GuestStandardAction, GuestState>(
                         record.modelId,
                         record.fieldId,
                         record.index,
-                        dragContext.dragged.path
+                        (dragContext.dragged as SearchItem).path
                       );
                     }
                     break;
@@ -833,7 +833,7 @@ const epic = combineEpics<GuestStandardAction, GuestStandardAction, GuestState>(
       ofType(assetDragStarted.type),
       withLatestFrom(state$),
       switchMap(([, state]) => {
-        if (nullOrUndefined(state.dragContext.dragged.path)) {
+        if (nullOrUndefined((state.dragContext.dragged as SearchItem).path)) {
           console.error('No path found for this drag asset.');
         } else {
           return initializeDragSubjects(state$);