+++ ui/guest/src/utils/util.ts
@@ -14,8 +14,10 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { ICERecord } from '../models/InContextEditing';
+import { ElementRecord, ICERecord } from '../models/InContextEditing';
 import { pluckProps } from '@craftercms/studio-ui/utils/object';
+import LookupTable from '@craftercms/studio-ui/src/models/LookupTable';
+import { ContentInstance, SandboxItem } from '@craftercms/studio-ui/models';
 
 export const foo = (...args: any[]) => void null;
 export const //
@@ -49,4 +51,22 @@ export function createLocationArgument() {
     'protocol',
     'search'
   );
+}
+
+export function isEditActionAvailable(args: {
+  record: ElementRecord | ICERecord;
+  models: LookupTable<ContentInstance>;
+  sandboxItemsByPath: LookupTable<SandboxItem>;
+  parentModelId: string | null;
+}): boolean {
+  const { record, models, sandboxItemsByPath, parentModelId } = args;
+  const model = models[record.modelId];
+  let path = model.craftercms.path;
+  if (!path) {
+    path = models[parentModelId].craftercms.path;
+  }
+  if (!path) {
+    console.error('* * * * *\nNo path found. Check.\n* * * * *', record, models);
+  }
+  return sandboxItemsByPath[path].availableActionsMap.edit;
 }